requirements.txt
asgiref==3.5.2
astroid==2.11.6
backports.zoneinfo==0.2.1
coverage==6.4.1
dill==0.3.5.1
Django==4.0.5
django-decouple==2.1
django-otp==1.1.3
djangorestframework==3.13.1
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
platformdirs==2.5.2
Pygments==2.12.0
pylint==2.14.3
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.11.0
typing-extensions==4.2.0
wrapt==1.14.1
